pyproject.toml
[project]
name = "fastapi-socialmedia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "fastapi[all]>=0.115.6",
    "psycopg>=3.2.4",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]
